tests: make a simple test suite via _quarto.yml
Unfortunately, there's no simple way to run tests to see if the documents in
this template are correctly rendered. To remedy this, I've created a set of
testing documents that can all be rendered together as a quarto project. I've
also added some bash scripts that will move the intermediate tex and bib files
generated by quarto into corresponding folders in test-results (TODO: replace
the sh files with something that's more compatible across platforms). It's my
hope that by doing this we can write simple tests for the LaTeX itself in the
future.