tests: correct rendering of REVTeX font commands
This test is a visual check that the font commands listed in the REVTeX author
guide (and the Command and Options Summary) all render correctly. The majority
of these are written as inline LaTeX rather than markdown; while this *works*,
content using them won't render correctly in other document formats such as
html. In the long run, the ideal solution will be to create a set of pandoc
filters that allow use of these font commands in a cross platform way. (That
said, this isn't necessary for the math fonts: pandoc can usually convert math
mode LaTeX to whatever the target format requires.)